requirements.txt
adafruit-circuitpython-si7021==4.1.6
paho-mqtt==1.6.1
bluepy==1.3.0
pyyaml==6.0